requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
psycopg2-binary==2.9.10
python-jose[cryptography]==3.5.0
requests==2.32.5
python-dotenv==1.1.1
email-validator==2.3.0
